fix(blog): stop 500s on ISR pages caused by headers() in root layout
The apex-only GA/Meta Pixel gate read the hostname via next/headers in
the root layout, which is a dynamic API. Every page under the layout
that opts into static/ISR rendering (blog detail, revalidate=3600)
threw DynamicServerError (digest DYNAMIC_SERVER_USAGE) and returned a
plain 500 in production. The gate now lives in a client component that
reads window.location.hostname after mount — the scripts were
afterInteractive anyway, so behavior is unchanged and the layout is
static-safe again.
